int to double for score

+++ app/src/main/java/Models/Beach.java
@@ -11,12 +11,12 @@ public class Beach implements Parcelable {
     private String date;
     private String day;
     private double waveSizeFt;
-    private int score;
+    private double score;
     private ArrayList<String> warnings;
 
 
     public Beach(String spotId, String spotName, String date, String day,
-                 double waveSizeFt, int score, ArrayList<String> warnings) {
+                 double waveSizeFt, double score, ArrayList<String> warnings) {
         this.spotId = spotId;
         this.spotName = spotName;
         this.date = date;
@@ -32,7 +32,7 @@ public class Beach implements Parcelable {
         this.date = in.readString();
         this.day = in.readString();
         this.waveSizeFt = in.readDouble();
-        this.score = in.readInt();
+        this.score = in.readDouble();
         this.warnings = in.readArrayList(String.class.getClassLoader());
     }
 
@@ -48,7 +48,7 @@ public class Beach implements Parcelable {
         dest.writeString(date);
         dest.writeString(day);
         dest.writeDouble(waveSizeFt);
-        dest.writeInt(score);
+        dest.writeDouble(score);
         dest.writeList(warnings);
     }
 
@@ -64,7 +64,7 @@ public class Beach implements Parcelable {
         }
     };
 
-    public int getScore() {
+    public double getScore() {
         return score;
     }
